fix(devtools): don't collide with the default runtime function
Currently, we don't inject `runtime.js` because we collide with the `runtime.js` of the host Angular application. This creates problems with non-Angular apps which don't use the same app bundling and the `backend.js` does not execute. This PR renames the runtime function to `___ngDevToolsRuntime`, which fixes the issue.

Additionally, since we support only Chrome this PR also drops differential bundling. We support the last two major versions of Chrome (look at `browserlist`), but if users require, we can easily extend the set of supported browsers.

Lastly, to have predictable file names, this PR also drops the hash suffix in the `shell-chrome` project.